pyproject.toml
[project]
name = "research-paper-rag"
version = "0.1.0"
description = "End-to-end RAG system for research papers with multimodal PDF parsing, parent-child chunking, hybrid retrieval, and Streamlit chatbot."
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.31.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-groq>=0.2.0",
    "langchain-experimental>=0.3.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "pymupdf4llm>=0.0.10",
    "easyocr>=1.7.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "pandas>=2.0.0",
]

[project.scripts]
ingest = "ingest:main"
